+++ SAMUEL/apps/api/src/modules/field/field.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
+import { RoutesModule } from '../routes/routes.module';
 import { FieldController } from './field.controller';
 import { FieldService } from './field.service';
 
 @Module({
-  imports: [AuthModule],
+  imports: [AuthModule, RoutesModule],
   controllers: [FieldController],
   providers: [FieldService],
 })